feat: refactor weather application structure, enhance user authentication, and improve UI interactions
