+++ admin.html
@@ -5,9 +5,7 @@
   <title>Admin - Quản lý thời tiết</title>
   <meta name="viewport" content="width=device-width, initial-scale=1" />
 
-  <!-- Google Fonts -->
   <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;600&display=swap" rel="stylesheet" />
-  <!-- Bootstrap 5 -->
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" />
     <link rel="stylesheet" href="css/admin.css">    
 </head>
@@ -61,9 +59,8 @@
     </div>
   </div>
 
-  <!-- Firebase JS -->
   <script type="module">
-    // Chuyển tab sidebar
+  
     document.querySelectorAll(".nav-link").forEach(link => {
       link.addEventListener("click", e => {
         e.preventDefault();
