+++ index.html
@@ -5,11 +5,11 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Dự báo thời tiết</title>
 
-    <!-- Bootstrap -->
+    <meta name="description" content="Dự báo thời tiết chính xác và nhanh chóng cho các thành phố trên toàn thế giới. Tìm kiếm thời tiết hiện tại, dự báo theo giờ và theo ngày." />
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" />
-    <!-- Google Font -->
     <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;600&display=swap" rel="stylesheet" />
     <link rel="stylesheet" href="css/style.css">
+
   </head>
 
   <body class="d-flex flex-column min-vh-100">
@@ -33,6 +33,7 @@
             <button class="btn btn-secondary dropdown-toggle" type="button" id="dropdownMenuButton"data-bs-toggle="dropdown" aria-expanded="false">
               <span id="userName">Người dùng</span>
             </button>
+            
             <ul class="dropdown-menu dropdown-menu-end" aria-labelledby="dropdownMenuButton">
                 <li><a class="dropdown-item" href="taikhoan.html">Tài khoản của tôi</a></li>
                 <li><hr class="dropdown-divider" /></li>
@@ -54,8 +55,8 @@
           <div class="col-md-10">
             <div class="input-group shadow-sm">
               <input type="text" id="timkiem" class="form-control" placeholder="Nhập tên thành phố..." />
-              <button class="btn btn-primary" onclick="TimKiem()">Tìm kiếm</button>
-              <button class="btn btn-outline-secondary" onclick="doiDonVi()">°C / °F</button>
+              <button class="btn btn-primary" id="btn-Search">Tìm kiếm</button>
+              <button class="btn btn-outline-secondary" id="btn-changeTemp">°C / °F</button>
             </div>
             <div id="error" class="text-danger mt-2"></div>
           </div>
@@ -72,7 +73,6 @@
 
         </div>
 
-        <!-- Hourly Forecast -->
         <div class="weather-card">
           <h5 class="mb-3">⏰ Dự báo theo giờ</h5>
           <div id="hourly-forecast"></div>
@@ -96,10 +96,9 @@
     <!-- Scripts -->
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-    <script src="js/script.js"></script>
-    <script src="js/UI.js"></script>
-    <script type="module" src="js/luudiadiem.js"></script>
-    <script type="module" src="js/auth.js"></script>
+  
+    <script type="module" src="js/script.js"></script>
     <script type="module" src="js/main.js"></script>
+    <script type="module" src="js/auth.js"></script>
   </body>
 </html>
