+++ list_savelocation.html
@@ -28,5 +28,5 @@
     </footer>
     
   </body>
-  <script type="module" src="js/lay_danhsach_diadiem.js"></script>
+  <script type="module" src="js/Location.js"></script>
 </html>
